added file
Dear Soledad,
This script compute the correct averages for each hour of the days, the script in computeAVG.py has an error, and skip
the first row for each day, check it carefully.
The new data must be correct. 

Moreover, in some case the values are negative, I asked to a collegue of mine if there is a meaning and they told me that
it is normal, in the night the panel act as a load to the circuit, i.e. it is like a resistance, that absorbe a very small current.